fix unit tests and lint
