+++ design/ui/pagination/src/Pagination.CustomTable.fixture.js
@@ -1,0 +1,50 @@
+import React from 'react';
+import pt from 'prop-types';
+import { Table } from 'semantic-ui-react';
+
+const propTypes = {
+  data: pt.arrayOf(
+    pt.shape({
+      avatar: pt.string,
+      email: pt.string,
+      first_name: pt.string,
+      id: pt.number,
+      job_title: pt.string,
+      last_name: pt.string,
+    })
+  ).isRequired,
+};
+
+/**
+ * An example on how to separate the content
+ * from the pagination.
+ */
+const CustomTable = ({ data }) => (
+  <Table celled padded style={{ paddingTop: '0' }}>
+    <Table.Header>
+      <Table.Row>
+        {Object.keys(data[0]).map((field) => (
+          <Table.HeaderCell key={field}>{field}</Table.HeaderCell>
+        ))}
+      </Table.Row>
+    </Table.Header>
+    <Table.Body>
+      {data.map((row) => (
+        <Table.Row key={row.id}>
+          {Object.keys(row).map((field) => (
+            <Table.Cell key={field}>
+              {field === 'avatar' ? (
+                <img alt='employee' src={row[field]} />
+              ) : (
+                row[field]
+              )}
+            </Table.Cell>
+          ))}
+        </Table.Row>
+      ))}
+    </Table.Body>
+  </Table>
+);
+CustomTable.propTypes = propTypes;
+
+export default CustomTable;
